requirements.txt
fastapi
uvicorn
faiss-cpu
psutil
sentence-transformers
